全部IMetaDrawableかIMetaAudiableで描写するようにした まだ未完成

+++ Metasia.Core/Objects/IMetaDrawable.cs
@@ -9,6 +9,6 @@ namespace Metasia.Core.Objects
 {
 	public interface IMetaDrawable
 	{
-		public void DrawExpresser(ref DrawExpresserArgs e, ulong frame);
+		public void DrawExpresser(ref DrawExpresserArgs e, int frame);
 	}
 }
